Refactor: Extract new mod `unix/linux.rs`
